파일 업로드
파일업로드 서비스(신규에 저장만) 작성 후 오류 수정중

+++ board/src/main/java/com/example/board/board/dto/BoardFileDto.java
@@ -8,5 +8,5 @@ public class BoardFileDto {
     private int boardIdx;
     private String originalFileName;
     private String storedFilePath;
-    private int fileSize;
+    private long fileSize;
 }
